EshInfCheck split int two checks
- Created abstract class;
- Created seperate classes for each check;
- Changed the severity level.

Closes #85
