+++ src/main/resources/rulesets/checkstyle/rules.xml
@@ -98,11 +98,15 @@
      <property name="ignoreImportedPackages" value="org.junit.*"/>
   </module>
    
-  <module name="org.openhab.tools.analysis.checkstyle.EshInfXmlCheck">
-    <property name="severity" value="warning" />
+  <module name="org.openhab.tools.analysis.checkstyle.EshInfXmlValidationCheck">
+    <property name="severity" value="error" />
     <property name="thingSchema" value="http://eclipse.org/smarthome/schemas/thing-description-1.0.0.xsd" />
     <property name="bindingSchema" value="http://www.eclipse.org/smarthome/schemas/binding-1.0.0.xsd" />
     <property name="configSchema" value="https://www.eclipse.org/smarthome/schemas/config-description-1.0.0.xsd" />
+  </module>
+  
+  <module name="org.openhab.tools.analysis.checkstyle.EshInfXmlUsageCheck">
+    <property name="severity" value="warning" />
   </module>
   
   <module name="org.openhab.tools.analysis.checkstyle.PomXmlCheck">
