+++ src/test/resources/checks/checkstyle/eshInfXmlUsageCheckTest/missingConfigDesciptionRef/ESH-INF/thing/thing-types.xml
