+++ src/test/resources/checks/checkstyle/eshInfXmlUsageCheckTest/missingConfigDesciptionRef/build.properties
