+++ src/test/resources/checks/checkstyle/eshInfXmlUsageCheckTest/missingSupportedBridgeRef/ESH-INF/thing/thing-types.xml
