+++ src/test/resources/checks/checkstyle/eshInfXmlUsageCheckTest/missingSupportedBridgeRef/build.properties
