+++ src/test/resources/checks/checkstyle/eshInfXmlUsageCheckTest/unusedBridge/ESH-INF/thing/thing-types.xml
