+++ src/test/resources/checks/checkstyle/eshInfXmlUsageCheckTest/unusedBridge/build.properties
