+++ src/test/resources/checks/checkstyle/eshInfXmlUsageCheckTest/unusedConfigDescription/build.properties
