+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/emptyThingTypeXml/build.properties
