+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/invalidBinding/ESH-INF/binding/bind.xml
