+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/invalidBinding/build.properties
