+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/invalidConfig/ESH-INF/config/conf.xml
