+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/invalidConfig/build.properties
