+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingChannelTypeContent/ESH-INF/thing/thing-types.xml
