+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingChannelTypeContent/build.properties
