+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingPropertyContent/ESH-INF/thing/thing-types.xml
