+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingPropertyContent/build.properties
