+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingThingDescriptionsContent/ESH-INF/thing/thing-types.xml
