+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingThingDescriptionsContent/build.properties
