+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingThingTypeContent/ESH-INF/thing/thing-types.xml
