+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingThingTypeContent/build.properties
