+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingXmlFileInBuildProperties/ESH-INF/thing/thing-types.xml
