+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/missingXmlFileInBuildProperties/build.properties
