+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/sequenceBridgeTypeCheck/ESH-INF/thing/thing-types.xml
