+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/sequenceBridgeTypeCheck/build.properties
