+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/sequenceChannelTypeCheck/ESH-INF/thing/thing-types.xml
