+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/sequenceChannelTypeCheck/build.properties
