+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/sequenceThingTypesCheck/ESH-INF/thing/thing-types.xml
