+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/sequenceThingTypesCheck/build.properties
