+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/thingTypesConfiguration/ESH-INF/binding/bind.xml
