+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/thingTypesConfiguration/ESH-INF/config/conf.xml
