+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/thingTypesConfiguration/ESH-INF/dummy/dummy.xml
