+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/thingTypesConfiguration/ESH-INF/thing/thing-types.xml
