+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/thingTypesConfiguration/build.properties
