+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/validBridgeType/ESH-INF/thing/thing-types.xml
