+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/validBridgeType/build.properties
