+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/validSupportedBridgeRef/ESH-INF/thing/bridge.xml
