+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/validSupportedBridgeRef/build.properties
