+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/validThingTypeXml/ESH-INF/thing/thing-types.xml
