+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/validThingTypeXml/build.properties
