+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/wrongDirectory/ESH-INF/thing-types.xml
