+++ src/test/resources/checks/checkstyle/eshInfXmlValidationCheckTest/wrongDirectory/build.properties
